local/requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.8
pycparser==2.19
python-telegram-bot==12.3.0
requests==2.22.0
requests-file==1.4.3
selenium==3.141.0
selenium-requests==1.3
six==1.13.0
soupsieve==1.9.5
tldextract==2.2.2
tornado==6.0.3
urllib3==1.25.7
